fix(apigateway): consolidate lambda permissions when reused for multiple operations
The maximum Lambda permission policy size could be exceeded for APIs which reused
the same Lambda function for multiple operations, as the integration added a new
permission for each operation, scoped down to the specific operation.

This change updates both the REST and HTTP API lambda integrations to consolidate
permissions when more than 10 permissions would be added for the same handler, creating a
permission scoped to the entire API rather than the operation. The behaviour
remains the same where individual lambdas are used for operations.

Fixes #9327
Fixes #19535