fix(lambda): can't find entry file under ESM module system (#35797)
### Issue # (if applicable)

Closes #21630.

### Reason for this change

In the ESM module system, callsites returns filenames prefixed with 'file://'. This is not compatible with the NodeJS file utility functions such as fs.existsSync().

### Description of changes

Remove 'file://' prefix.

### Describe any new or updated permissions being added

No new IAM permissions are added.

### Description of how you validated changes

Unit tests. Integration test cannot be added as the integration test system uses CommonJS, it will be impossible to replicate the error which only happens under an ESM system. 

### Credits

This is mostly inspired by @okko for [his PR](https://github.com/aws/aws-cdk/pull/21802). He found the root cause and fix for this issue, but it didn't get merged due to a lack of integration tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*