fix(s3-deployment): handle empty string in Source.data()  (#35824)
### Issue # (if applicable)

Closes #35809.
Original PR: https://github.com/aws/aws-cdk/pull/35818
This PR is been created to fasten release of the fix for the p0

### Reason for this change


`Source.data()` fails with a TypeError when called with an empty string as the data parameter in CDK v2.207.0+. This is a regression from v2.206.0 where empty string deployment worked correctly.

The issue manifests as:
```
TypeError: The "data" argument must be of type string or an instance of Buffer, TypedArray, or DataView. Received undefined
    at Object.writeFileSync (node:fs:2434:5)
```

### Description of changes
This change handles the the case from renderData function.


### Describe any new or updated permissions being added




### Description of how you validated changes
Unit tests have been added to cover this case and integ test has also been updated.



### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*