docs(eks-v2-alpha): clarify v2 capacity architecture (#35677)
### Issue # (if applicable)

Closes #<issue number here>.

### Reason for this change

The readme was incorrectly showing that the default capacity is managed node group, but this is not the case in v2. in V2 the default is EKS auto mode, so the readme was updated to show the current default as the architecture, with some clarification on the other possible capacity modes.

### Description of changes

Updated the readme to reflect the state of eks v2.

### Describe any new or updated permissions being added

None

### Description of how you validated changes

None

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*