fix(ecs): remove empty CfnClusterCapacityProviderAssociations resource (#35783)
### Issue # (if applicable)

Closes #35699
Related to: #35742

### Reason for this change

CfnClusterCapacityProviderAssociations is not needed when using only managed instances as a capacity provider, but it was being created anyway with empty values. It is needed if default capacity providers are set, or the capacity provider is not of the managed instances type.

### Description of changes

Removed the `clusterScopedCapacityProviderNames` check when creating CfnClusterCapacityProviderAssociations.

Destructive changes: CfnClusterCapacityProviderAssociations will be deleted if user does not provide default capacity providers or default fargate and uses only ManagedInstancesCapacityProvider.


### Describe any new or updated permissions being added

None

### Description of how you validated changes



### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*