chore: reapply refactoring for L1 resource relationships
This applies the code for L1 resource relationships, originally applied
in #35713 and reverted in #35832.

The reason is that another PR is also making changes in this area,
branched in between the original application and its revert, and is now
experiencing severe merge conflicts.

In this PR, reintroduce the large refactorings made in the original PR,
but disable their application via a boolean flag.