chore: reapply refactoring for L1 resource relationships (#35884)
This applies the code for L1 resource relationships, originally applied in #35713 and reverted in #35832.

The reason is that another PR is also making changes in this area, branched in between the original application and its revert, and is now experiencing severe merge conflicts.

In this PR, reintroduce the large refactorings made in the original PR, but disable their application by making the list of services we apply the relationships to empty.

```ts
export const RELATIONSHIP_SERVICES: string[] = [];
```

There were some changes to L2s in the original PR that are not included in this PR; they will need to be reapplied when restoring the feature.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*