fix: onEvent function to pass all the options to rule resource (#35829)
### Issue # (if applicable)

N/A

### Reason for this change

`onEvent()` has a lot of options other than description e.g. `ruleName` `detail` etc..
The old code didn't pass those values to rule, so it will never be propagated properly, so the consumer of this method will think that we're respecting their options, but in real we're discarding it.

### Description of changes

Creating the rule first with all the options so all the options from the consumer get propagated properly to the rule, then add the specific things that you want.

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

Add unit tests

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*