chore(spec2cdk): refactor to make it easier to extend
The `spec2cdk` tool was originally written to generate every service
submodule insularly.

We want to now extend it to emit structures into a new shared submodule.
In order to do that, this refactors the generator to have a more shared,
mutable object-oriented design which will be easier to share state in.

In addition, simplify the API around file name patterns, because they
were being overly flexible and annoying to work with for no real
benefit.