chore(spec2cdk): refactor to make it easier to extend (#35950)
The `spec2cdk` tool was originally written to generate every service submodule insularly.

We want to now extend it to emit structures into a new shared submodule. In order to do that, this refactors the generator to have a more shared, mutable object-oriented design which will be easier to share state in.

The core of the change looks like this:

```ts
// Old
for (const service of services) {
  const module = AstBuilder.forService(service);

  const writer = new TsFileWriter();
  writer.write(module);
}

// New
const ast = new AstBuilder();
for (const service of service) {
  ast.addService(service);
}

const writer = new TsFileWriter();
ast.writeAll(writer);
```

* Make the API based more around mutable objects with shared state.
* Remove the specific `ServiceModule`, `ResourceModule` classes, because they were adding no benefit over the generic `Module`.
* Simplify the API around file name patterns, because they were being overly flexible and annoying to work with for no real benefit.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*