ci(integ-test-deployment): use integ runner with strict flag (#35974)
### Reason for this change
Previously `integration-test-deployment` was passing in silent if some of the test were not found. 
We want to make sure that all the changed integ tests in a PR ran, or workflow should fail.

### Change
`integration-test-deployment` to use `integ-runner` with `--strict` flag so that workflow fails in case of test not found.


### Testing
Test run on fork: https://github.com/Abogical/aws-cdk/actions/runs/19229944532/job/54965890882?pr=19
### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*