feat(apigateway): add option for consolidating lambda permissions for rest and http lambda integrations
The maximum Lambda permission policy size can be exceeded for APIs which reuse
the same Lambda function for multiple operations, as the integration adds a new
permission for each operation, scoped down to the specific operation.

This change updates both the REST and HTTP API lambda integrations with options
to scope the permission to any operation on the API, adding a single statement
and avoiding overflowing the maximum policy size.

Fixes #9327
Fixes #19535