fix: automatically clean up unit test assemblies
The CDK framework creates a temporary directory in `/tmp` for every unit
test, and never cleans those up.

In this PR, register all temporary assembly directories created by doing
`new Stack()` or `new App()` without an `outdir`, and delete them when
the Node process exits.

This will only affect unit tests: if the `outdir` property is set
explicitly, or the CDK App is being synthesized by the CLI (and
`$CDK_OUTDIR` is set), the assembly directory will not be cleaned.