fix(s3-deployment): `Source.jsonData()` fails with null JSON values (#36054)
In cdk-lib aws-s3-deployment, creating a Source.jsonData() will fail if the JSON has any null fields in it due to `escapeTokens` trying to build an object from it, resulting in `TypeError: Cannot convert undefined or null to object`

### Issue # (if applicable)

Closes #36052.

### Reason for this change

Can't create a Source.jsonData() when `null` values are present in the JSON data itself

### Description of changes

Added a null check around `escapeTokens()` to prevent it from trying to process `null` values

### Describe any new or updated permissions being added

N/A


### Description of how you validated changes

Tested by hand via modifying the compiled JS in my project to add the check, and was able to successfully upload a JSON file to S3 with a null value contained inside. Also added coverage via unit tests/integration tests for the PR build to chew through.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*