fix(aws-bedrock-agentcore-alpha): empty submodule accidentally exposed
`perms.ts` is accidentally being exported. It only contains a namespace
with static const arrays of strings, which are of no use to jsii
consumers. These represent IAM actions that are being set inside the
library, and don't need to be public to begin with.

Move the file into a directory called `private` to make it clear
it should not be exported, and remove it from the jsii entry point.