chore(mixins-preview): improved find bucket policy (#36144)
### Reason for this change

A helper function to reflect on the construct tree to find the closet Bucket Policy for a Bucket.

### Description of changes

Implemented a generic function to find two related resources that are "close" to each other. Close is defined here as:
- the closest (transitive) child matching the predicate
- the above to the closest parent

Than added a specific function just for bucket. For now this is hand-written and not publicly exposed. If we can identify these kind of connection patterns, we can likely code gen this for all resources.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Unit tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*