fix(cloudformation-include): `TypeError` when including template with intrinsic functions (#36157)
### Issue # (if applicable)

Closes #36140

### Reason for this change

In #35838 we introduced typed relationships for L1 constructs. For properties with arrays, this incorrectly assumed that the value could not be a token. But it can and especially when using the `cloudformation-include` module.

### Description of changes

Adds a check to ensure the value is not a Token, which cannot be used as an array, despite passing the type check.

There are two places in `resolver-builder.ts` where we added a `.forEach()`. The first one is fixed in this PR, second location already has a check for a resolvable.

Note we are only checking for a resolvable here, not if the value is actually an array. At some point this could extended to also check for an array, likely with some refactoring of the "in-place" mapping into a helper function (the currently generated code is very unreadable). For this PR I have chosen to do the minimal change.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Added a new unit test for this case.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*