feat: L2 `EventPattern` can be used for `CfnRule`
It is now possible to use the L2 `EventPattern` interface to synthesize
events that will work with the L1 `CfnRule` class:

```ts
// L1
interface CfnRuleProps {
  readonly eventPattern: any;            // <-- not typed
}

// L2
interface RuleProps {
  readonly eventPattern: EventPattern;   // <-- typed
}
```

At the L1, there is "raw data"; the L2 translates the fields of
the `EventPattern` type to the correct fields of "raw data".

In practice, the only difference is that one of the raw data
fields is called `detail-type`, which in the L2 type is called
`detailType`. All other fields are the same.

We add the ability to render specific fields of specific L1s
differently if we so choose: the renderer for `CfnRule` now
has gotten the ability to specifically recognize `detailType`
and translate it to `detail-type`.

This now makes both classes take the same inputs.