feat(events): the L2 `EventPattern` interfaces can be used with `CfnRule` (#36191)
Objects that conform to the L2 `EventPattern` interface are now interpreted correctly by the L1 `CfnRule` class:

```ts
// L1
interface CfnRuleProps {
  readonly eventPattern: any;            // <-- not typed
}

// L2
interface RuleProps {
  readonly eventPattern: EventPattern;   // <-- typed
}
```

At the L1, there is "raw data"; the L2 translates the fields of the `EventPattern` type to the correct fields of "raw data".

In practice, the only difference is that one of the raw data fields is called `detail-type`, which in the L2 type is called `detailType`. All other fields are the same.

We add the ability to render specific fields of specific L1s differently if we so choose: the renderer for `CfnRule` now has gotten the ability to specifically recognize `detailType` and translate it to `detail-type`.

This now makes both classes take the same inputs, reducing a common cause of user confusion and mistakes.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*