fix(ecs): wrong ARN generated in `Cluster.grantTaskProtection` method (#36207)
The `Cluster.grantTaskProtection()` method is delegating to the method in the auto-generated `ClusterGrants` class. This class is generating policies with an incorrect ARN. It's using the cluster ARN (`:cluster/abc/*`) instead of the tasks ARN (`:task/abc/*`).

Since the code generation mechanism doesn't provide any way of specifying a custom logic for the ARN generation, replace the auto-generated `ClusterGrants` with hand-written one, in which the ARN is computed by:

```ts
  private arnForTasks(keyPattern: string): string {
    return Stack.of(this.resource).formatArn({
      service: 'ecs',
      resource: 'task',
      resourceName: `${this.resource.clusterRef.clusterName}/${keyPattern}`,
      arnFormat: ArnFormat.SLASH_RESOURCE_NAME,
    });
  }
```

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*