feat(aws-cdk-lib): add `arnFor<ResourceName>` for 47 more resources (#36231)
### Reason for this change

Today, out of `1512` resources, `892` have an `arnFor${Resource}` getter function (~58%).
The reasons are varied, but it comes done to us not being able to identify the required resource properties in an automated way.

### Description of changes

We are adding a few additional heuristics to the matching:

- If we cannot find an Arn property for a resource, but can match all variables of its arn template to properties, we can use that to interpolate the Arn . (`37` additional resources) 
- A property may now also be called or end in `ARN` (all caps). (`9` additional resources) 
- Some resources use a property name prefix that is different than their resource name, e.g. [AWS::BedrockAgentCore::Runtime](https://docs.aws.amazon.com/AWSCloudFormation/latest/TemplateReference/aws-resource-bedrockagentcore-runtime.html) uses `AgentRuntime` instead of the expected `Runtime`. We extract this prefix from a primary identifier if it is ending in `Id`. (`2` additional resources) 

This brings our total count to `939` or `~62%`. A little better!

PR also includes a major refactor of how reference props and now arn getters are organized. A lot of code depends on this information, so best to centralize the logic for it.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

Existing and additional tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*