chore: cannot write file '.../aws-iam/index.d.ts' because it would overwrite input file
We have been starting to get the above error, because of the style
of imports we've been using.

If *inside* `aws-cdk-lib` we use an import of the form:

```ts
import * as iam from 'aws-cdk-lib/aws-iam';
```

Then we add an `index.d.ts` file to the set of inputs, despite there
being an `index.ts` right next to it.

Undo that everywhere, and add a linter rule for it.