chore: cannot write file 'aws-iam/index.d.ts' because it would overwrite input file (#36262)
We have been starting to get the above error, because of the style of imports we've been using.

If *inside* `aws-cdk-lib` we use an import of the form:

```ts
import * as iam from 'aws-cdk-lib/aws-iam';
```

Then we add an `index.d.ts` file to the set of inputs, despite there being an `index.ts` right next to it.

This code was being generated by the Grants module generator. Fix it to use a relative import.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*