fix(mixins-preview): service exports are different then in `aws-cdk-lib` (#36201)
### Issue

Fixes #36210

### Reason for this change

`@aws-cdk/mixins-preview` exports services as `aws_s3` (note the underscore).
But `aws-cdk-lib` exports them as `aws-s3` (note the dash).

The intention was to align how the exports are named for ease of future migration. Better to fix this now.

### Description of changes

Renaming `aws_<SERVICE>` to `aws-<service>` and the required update of `cdk-generate-synthetic-examples`.

### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

test pipeline

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*