feat(lambda): add support for durable functions (#36282)
### Issue # (if applicable)

N/A

### Reason for this change
AWS Lambda L2 support for the new durable functions feature.

https://docs.aws.amazon.com/lambda/latest/dg/durable-functions.html

### Description of changes
Updating the `Lambda::Function` L2 construct with the new `durableConfig` property.

### Describe any new or updated permissions being added
Updating the managed policy that gets attached to lambda functions to be the new `AWSLambdaBasicDurableExecutionRolePolicy` managed policy for durable functions. 

In addition to the existing policies that the `AWSLambdaBasicExecutionRole` has, this new policy has permissions for `lambda:CheckpointDurableExecution` and `lambda:GetDurableExecutionState` actions.

### Description of how you validated changes
Unit tests, integration tests and generated snapshot files.

### Checklist
- [X] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*