chore: fix Docker tests on SE Linux (#36284)
On SE Linux, apparently we are mounting volumes with `z,delegated` while on other machines we are just using the string `delegated`.

Our tests are all written to expect the string `delegated`, which means they fail when run on an SE Linux machine.

Update the tests to mirror the real behavior exactly.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*