chore: fix codegen for new service modules (#36296)
### Reason for this change

Recent refactors introduced a bug for the codegen for new service modules. The service module `index.ts` file would be generated with the wrong code path.

### Description of changes

Fix the computation of the export path for new service modules.

During debugging I also noticed that the temporary schema override did not work anymore. So this was fixed as well.
I've also deleted the one temporary schema that was currently in the repo. It is now redundant as the changes have been made available upstream.


### Describe any new or updated permissions being added

n/a

### Description of how you validated changes

existing tests plus manual test of dropping the schemas in question into the temporary schema folder

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*