fix: re-export of `ResourceEnvironment` is not an alias
We changed the FQNs of interfaces in a recent PR in a non-backwards
compatible way. By keeping aliases for the old types in TypeScript this
should keep automatic backwards compatibility for TypeScript users,
which translates to backwards compatibility of Java users of
TypeScript-authored libraries.

However, we made a mistake in the aliasing. Originally, we wrote:

```ts
export type { IEnvironmentAware, ResourceEnvironment } from '...';
```

But `jsii` treats `export ... from` specially: it will take elements
from the given file and export them *as if* they had been written
in the current file (= current submodule).

This led to 2 equivalent but different definitions of `ResourceEnvironment`:

```
aws-cdk-lib.interfaces.ResourceEnvironment   (intended)
aws-cdk-lib.ResourceEnvironment              (unintended)
```

Change the syntax to:

```ts
import { IEnvironmentAware, ResourceEnvironment } from '...';
export type { IEnvironmentAware, ResourceEnvironment };
```

`export ...` is not the same as `export ... from`, so this now
reduces us back to one definition with a TypeScript alias.