chore: aws-cdk-lib tests were not being typechecked (#36294)
In order to support some other performance-related changes, I was planning to split the sets of source files inside `aws-cdk-lib` into 2 classes:

- "Library" source files: all `.ts` files except those in `test` subdirectories.
- "Test" source files: files in `test` subdirectories, consuming the library as a separately compiled subproject.

It turns out this does not work for 2 reasons: 

- `stripInternal: true` (TypeScript feature): there are some APIs that we strip out of the `.d.ts` files. If we put a discrete compilation step between the library and the tests, that means that tests can't use `@internal` APIs either, which we test in a number of cases.
- `--strip-deprecated` (jsii feature): there are some legacy `@deprecated` APIs that we strip out of the `.d.ts` files as well, and the same applies: we have tests for a number of these that start failing once we separately compile the two source sets.

Instead, we're just absorbing the test suite into a single `tsconfig.json` for this project again.

The upshot of doing all this is: **test files are now being type-checked by `tsc` and it turns out they had a ton of type errors in them!**

In most cases we saw "unused variable" errors, but there were honest to goodness typing mistakes in there, that just happened to execute properly when runing through `ts-jest` (with type checking ignored).

I'm submitting this as a PR by itself to get those changes out of the way.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*