chore: update PR build workflow to Node.js 20 (#36391)
### Issue # (if applicable)

N/A

### Reason for this change

The AWS CDK repository currently uses Node.js 18 in the PR build workflow, but Node.js 18 has reached end-of-life. Node.js 20 is now the recommended LTS version and should be used for CI/CD pipelines to ensure security updates and compatibility with modern tooling. See our [deprecation announcement](https://aws.amazon.com/blogs/devops/announcing-the-end-of-support-for-node-js-18-x-in-aws-cdk/).

### Description of changes

This change updates the PR build workflows to use Node.js 20 instead of Node.js 18, aligning with current best practices and ensuring the build environment stays up-to-date with supported Node.js versions.

### Describe any new or updated permissions being added

No new or updated permissions are required for this change.

### Description of how you validated changes

The change only affects the Node.js version used in GitHub Actions workflows. The workflow configuration has been updated and will be validated when the PR build runs.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*